% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4rexamples-package.R
\name{cpp4rexamples-package}
\alias{cpp4rexamples-package}
\title{Simple Examples for Demonstration Purposes}
\description{
Simple examples to showcase 'cpp4r' syntax. The example
 functions are quite simple (e.g., sum and square of a number) and with the
 goal to show what 'cpp4r' adds to usual 'C++' syntax.
}

