% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{sign_cpp}
\alias{sign_cpp}
\title{Sign of a number}
\usage{
sign_cpp(x)
}

\arguments{
\item{x}{A numeric value.}
}

\value{
An integer indicating the sign of the number.
}

\description{
Returns 1 if the number is positive, -1 if negative, and 0 if zero.
}

\examples{
sign_cpp(10)
sign_cpp(-5)
sign_cpp(0)
}

