% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{square_coordinates}
\alias{square_coordinates}
\title{Square Coordinates (C++)}
\usage{
square_coordinates(x)
}

\arguments{
\item{x}{integer vector}
}

\description{
It squares each element of an integer vector.
}

\examples{
square_coordinates(1:3)
}

