% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign.R
\name{sign_r}
\alias{sign_r}
\title{Sign of a number}
\usage{
sign_r(x)
}
\arguments{
\item{x}{A numeric value.}
}
\value{
An integer indicating the sign of the number.
}
\description{
Returns 1 if the number is positive, -1 if negative, and 0 if zero.
}
\examples{
sign_r(10)
sign_r(-5)
sign_r(0)
}
