% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{square_coordinates_2}
\alias{square_coordinates_2}
\title{Alternative Square Coordinates (C++)}
\usage{
square_coordinates_2(x)
}
\arguments{
\item{x}{integer vector}
}
\description{
It squares each element of an integer vector.
}
\examples{
square_coordinates_2(c(1, 2, 3))
}
